+++ src/components/AddCard/AddCardContainer/AddCardContainer.js
@@ -5,13 +5,18 @@ import { Title } from "../../style/layout";
 import useCardNumberFields from "./hooks/useCardNumberFields";
 
 const AddCardContainer = () => {
-  const { fields, handleChange } = useCardNumberFields();
+  const { fields, handleChangeSingleInput, handleChangeMultipleInput } =
+    useCardNumberFields();
 
   return (
     <>
       <Title>카드 추가</Title>
       <main>
-        <AddCardForm fields={fields} onChange={handleChange} />
+        <AddCardForm
+          fields={fields}
+          onChangeSingleInput={handleChangeSingleInput}
+          handleChangeMultipleInput={handleChangeMultipleInput}
+        />
       </main>
     </>
   );
