+++ src/components/AddCard/CardNumberInput/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CardNumberInput";
