+++ src/components/AddCard/InputBox/index.js
@@ -1,0 +1,1 @@
+export { default } from "./InputWithLabel";
