+++ src/components/AddCard/utils/isFullField.js
@@ -1,0 +1,5 @@
+import { MAX_LENGTH } from "../constants";
+
+const isFullField = (filed) => filed.length === MAX_LENGTH;
+
+export default isFullField;
